Challenge8<Password generator>.py
This program generates a random password of specified length 'passlen' using unique set of characters from a defined set  'x'. 

Here, the code imports the random module, specifies the character set, sets the password length, and uses the function 'random.sample()' to select a unique set of characters from the set 'x', and joins them into a single string.

#Generating unique characters
 'random.sample(x, passlen)' generates unique set of characters from the set 'x' without any repetition.

#Concatenating into a string
join() is an in-built string function which concatenates the elements of the list, (here, elements of x), thereby converting the list into a string.

#Use of  "" before 'join()'
"" before join() function is used as a separator to ensure that only characters from the given list is used to concatenate into a string, i.e., it avoids spaces between the joined characters.