Challenge11.py
//PALINDROME//

This Python code checks whether a given word(string) is a palindrome. 
A palindrome is a word that reads the same backward as forward.
The string is reversed using a method called 'string slicing', and is compared with the input string. If the reverse is equal to the input string, the strings are palindrome, else, they are not.

Here, we get the reverse of the string using a slice operator ':', is used to extract a portion of a sequence by specifying a range of indices.
Syntax: S[ start : stop : step ]
If S is a string, the expression S[ start : stop : step ] returns the portion of the string from index start to index stop, at a step size step.
Here, we are reversing the string using a step of '-1'.