Challenge15.py
The code reads an input from the user and prints consecutive numbers from 1 to n without adding spaces or newlines in between them.

Here, the if __name__ == '__main__': construct is used in Python to allow or prevent certain code from being run when the script is executed directly versus when it is imported as a module into another script. That is, this line ensures that the code block underneath it runs only if the script is executed directly (not imported as a module).

The 'for' loop iterates over a range of numbers from 1 to n. 
The range function generates numbers starting from 1 up to, but not including, n + 1.