Challenge16.py
In this code, the 'print_full_name' function takes two parameters: 'first' and 'last', which are strings representing a person's first and last names.
This function concatenates the 'first' and 'last' names with a space in between, thereby, giving us the function 'full_name'. This concatenation is done using an f-string, that allows you to embed expressions inside string literals, using curly braces {}.
The if __name__ == '__main__': block ensures that the code only executes when the script is run directly. Now, within this block, the code prompts the user to input their first and last names.

It then calls the 'print_full_name' function with the provided names, displaying the final greeting message. :)