Challenge20.py
                                //SIZE OF TUPLES//
The program demonstrates how to calculate the size of different tuples using the __sizeof__() method.

Tuple1: Contains a mix of integers and strings. 
The code calculates and prints the size of Tuple1 in bytes using Tuple1.__sizeof__().

Tuple2: Contains only integers. 
Similarly, it calculates and prints the size of Tuple2 in bytes.

Tuple3: Contains tuples of pairs, each consisting of an integer and a string. 
Again, it calculates and prints the size of Tuple3 in bytes.

The code utilizes 'sys' module to print the size of each tuple. 
This can be useful for understanding memory usage when working with tuples in Python. 
☜(ﾟヮﾟ☜)